# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NORA API
  version: 0.2.0
  description: |
    API REST do backend NORA (Spring Boot 3 / Java 21). Cobre autenticacao com cookies httpOnly,
    IAM estilo AWS (grupos + policies + attachments + audit), convites por e-mail, contexto comercial
    do tenant, dominio corporativo, upload/listagem/detalhe de reunioes, gestao de tarefas extraidas,
    analise ao vivo e emissao de token efemero para Azure Speech.

    ## Autenticacao
    A maioria dos endpoints exige um JWT. Ha dois modos de transporte equivalentes:

    - **Bearer** — header `Authorization: Bearer <accessToken>` (esquema `bearerAuth`).
    - **Cookie httpOnly** — `POST /auth/login` (e `POST /iam/invites/{token}/accept`) define
      `Set-Cookie: nora_access` (`Path=/`, `SameSite=Lax`, `HttpOnly`) e `Set-Cookie: nora_refresh`
      (`Path=/auth`, `SameSite=Strict`, `HttpOnly`). O navegador reenvia `nora_access` automaticamente.
      `nora_refresh` so chega nas rotas sob `/auth` (refresh/logout). A flag `Secure` e ligada em prod
      via `nora.auth.cookie-secure`.

    O campo `accessToken` continua presente no corpo JSON do login por compat com clientes legados;
    novos clientes devem confiar nos cookies.

    ## Autorizacao IAM (ADR 0007)
    NAO existem papeis fixos. A autorizacao e baseada em **policies IAM** (estilo AWS) avaliadas por
    `(action, resource)`. Cada endpoint protegido declara na descricao a action IAM exigida
...
    `403 IAM_FORBIDDEN`.

    ## Formato de erro
...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /iam/users/invite:
    post:
      tags: [Invitations]
      summary: Convida um usuario por e-mail
      description: |
        Action IAM: `iam:user:invite`. O `token` do convite NUNCA e devolvido na resposta. E-mail
        fora do dominio permitido retorna `422 EMAIL_DOMAIN_NOT_ALLOWED`; convite pendente duplicado
        `409 INVITE_DUPLICATE_PENDING`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InviteUserRequest' }
      responses:
        '201':
          description: Convite criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invitation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /iam/invites/{token}/accept:
    post:
      tags: [Invitations]
      security: []
      summary: Aceita um convite e cria a conta
      description: |
        Publico — o `token` da URL e a credencial. Cria o usuario, anexa aos grupos do convite e
        autentica: retorna `LoginResponse` e define os cookies `nora_access` e `nora_refresh`
        (igual ao login). Convite inexistente `404 INVITE_NOT_FOUND`, ja aceito
        `409 INVITE_ALREADY_ACCEPTED`, expirado `410 INVITE_EXPIRED`.
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcceptInviteRequest' }
      responses:
        '200':
          description: Conta criada e autenticada. Define cookies `nora_access` e `nora_refresh`.
          headers:
            Set-Cookie:
              description: Cookies httpOnly `nora_access` e `nora_refresh`.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410': { $ref: '#/components/responses/Gone' }
  /iam/invites:
    get:
      tags: [Invitations]
      summary: Lista convites do tenant
      description: 'Action IAM: `iam:invite:read`. Filtro opcional por status. Paginacao no MVP e client-side.'
      parameters:
        - in: query
          name: status
          required: false
          schema: { $ref: '#/components/schemas/InvitationStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InviteListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /iam/invites/{id}:
    delete:
      tags: [Invitations]
      summary: Revoga um convite pendente
      description: 'Action IAM: `iam:invite:revoke`.'
      parameters: [{ $ref: '#/components/parameters/InviteId' }]
      responses:
        '204': { description: Revogado }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----------------------------------------------------------------------------
  # Tenant — commercial context + domain
  # ----------------------------------------------------------------------------
  /tenant/context:
    get:
      tags: [Tenant]
      summary: Obtem o contexto comercial/produto do tenant
      description: |
        Action IAM: `tenant:context:read`. 1-1 com o tenant do JWT (sem path param). Ausente retorna
        `404 TENANT_CONTEXT_NOT_FOUND`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantContext' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Tenant]
      summary: Cria ou atualiza o contexto comercial/produto do tenant
      description: |
        Action IAM: `tenant:context:write`. Upsert do contexto injetado no prompt da LLM
        (companyName, valueProposition, products, objectionHandling).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantContextInput' }
      responses:
        '200':
          description: Salvo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantContext' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tenant/domain:
    get:
      tags: [Tenant]
      summary: Obtem o dominio de e-mail permitido do tenant
      description: 'Action IAM: `tenant:domain:read`.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantDomain' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      tags: [Tenant]
      summary: Atualiza o dominio de e-mail permitido do tenant
      description: |
        Action IAM: `tenant:domain:write`. `allowedEmailDomain` pode ser `null` para remover a
        restricao. Dominio invalido retorna `422 TENANT_DOMAIN_INVALID`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantDomainUpdateRequest' }
      responses:
        '200':
          description: Atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantDomainUpdateResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }

  # ----------------------------------------------------------------------------
  # Meetings
  # ----------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: nora_access
      description: |
        Cookie httpOnly definido por `POST /auth/login` (e por `POST /iam/invites/{token}/accept`).
        Equivalente ao Bearer; o navegador o reenvia automaticamente. Renovado via `POST /auth/refresh`.
  parameters:
...
    InviteId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Requisicao invalida (`VALIDATION_FAILED`, `MALFORMED_REQUEST`, `TOKEN_INVALID`, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Nao autenticado (`UNAUTHENTICATED`, `INVALID_CREDENTIALS`, `REFRESH_TOKEN_INVALID`, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Sem permissao (`FORBIDDEN`, `IAM_FORBIDDEN`, `USER_DISABLED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflito (`EMAIL_ALREADY_TAKEN`, `IAM_NAME_TAKEN`, `INVITE_ALREADY_ACCEPTED`, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Gone:
      description: Recurso expirado (`INVITE_EXPIRED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    UnprocessableEntity:
      description: Entidade nao processavel (`EMAIL_DOMAIN_NOT_ALLOWED`, `TENANT_DOMAIN_INVALID`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, traceId, timestamp]
      properties:
        code: { type: string, example: MEETING_NOT_FOUND }
        message: { type: string }
        traceId: { type: string }
        timestamp: { type: string, format: date-time }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }

    # ---- shared enums ----
...
    InvitationStatus:
      type: string
      enum: [PENDING, ACCEPTED, EXPIRED, REVOKED]

    # ---- auth ----
...
    LoginResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresInSeconds, userId, tenantId, email, displayName]
      description: |
        Corpo do login (e do aceite de convite). Tambem acompanha headers `Set-Cookie`
        (`nora_access`, `nora_refresh`). `accessToken`/`refreshToken` no corpo existem por compat
        legada; novos clientes devem confiar nos cookies httpOnly.
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresInSeconds: { type: integer, format: int64, description: TTL do access token em segundos. }
        userId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
...
    InviteUserRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        groupIds:
          type: array
          items: { type: string, format: uuid }
        expiresInDays: { type: [integer, 'null'] }
    AcceptInviteRequest:
      type: object
      required: [password]
      properties:
        displayName:
          type: string
          maxLength: 120
          description: Opcional; se omitido usa o local-part do e-mail.
        password: { type: string, minLength: 10, maxLength: 128 }
    Invitation:
      type: object
      required: [id, tenantId, email, status, invitedBy, invitedAt, expiresAt, groupIds]
      description: O token do convite NUNCA e devolvido.
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        email: { type: string, format: email }
        status: { $ref: '#/components/schemas/InvitationStatus' }
        invitedBy: { type: string, format: uuid }
        invitedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        groupIds:
          type: array
          items: { type: string, format: uuid }
        acceptedAt: { type: [string, 'null'], format: date-time }
        acceptedUserId: { type: [string, 'null'], format: uuid }
    InviteListResponse:
      type: object
      required: [items, total, page, size]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Invitation' }
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }

    # ---- tenant context ----
    TenantContextInput:
      type: object
      required: [companyName]
      properties:
        companyName: { type: string, maxLength: 200 }
        industry: { type: string, maxLength: 200 }
        valueProposition: { type: string, maxLength: 2000 }
        idealCustomerProfile: { type: string, maxLength: 2000 }
        products:
          type: array
          items: { $ref: '#/components/schemas/Product' }
        competitors:
          type: array
          items: { type: string, maxLength: 200 }
        objectionHandling:
          type: array
          items: { type: string, maxLength: 500 }
    TenantContext:
      type: object
      required: [tenantId, companyName]
      properties:
        tenantId: { type: string, format: uuid }
        companyName: { type: string }
        industry: { type: [string, 'null'] }
        valueProposition: { type: [string, 'null'] }
        idealCustomerProfile: { type: [string, 'null'] }
        products:
          type: array
          items: { $ref: '#/components/schemas/Product' }
        competitors:
          type: array
          items: { type: string }
        objectionHandling:
          type: array
          items: { type: string }
        updatedAt: { type: string, format: date-time }
    Product:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 1000 }
        keyDifferentiators:
          type: array
          items: { type: string, maxLength: 200 }

    # ---- tenant domain ----
    TenantDomain:
      type: object
      required: [tenantId]
      properties:
        tenantId: { type: string, format: uuid }
        allowedEmailDomain:
          type: [string, 'null']
          description: '`null` quando nao ha restricao de dominio.'
    TenantDomainUpdateRequest:
      type: object
      properties:
        allowedEmailDomain:
          type: [string, 'null']
          maxLength: 255
          description: '`null` para remover a restricao.'
    TenantDomainUpdateResponse:
      type: object
      required: [tenantId, updatedAt, updatedBy]
      properties:
        tenantId: { type: string, format: uuid }
        allowedEmailDomain: { type: [string, 'null'] }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string, format: uuid }

    # ---- meetings ----
...